Fix DispatcherThread; Use "Circuit Avionics" instead of master switch
This is much more authoritative, and easier for us to deal with than
trying to check all the ways it might lose power. I'm very impressed
and excited that this level of detail is available! :)

One bug with polling the vars, however, is that it introduces a
race condition between us setting the var and us receiving a
(potentially) stale value from the server....