Implement most of the basic Transponder UI
Ideally, we should still fit an "ident" button and the mode
switch, though the switch will probably be a pain. We will
probably have to subclass Button to get the numbers to
draw properly, since the text is cut off right now. Also nice
would be buttons for VFR and "clear," though with the nice touch
interface it's not that big a deal.