Disable the IDENT button; add @Named to transponder subject
Apparently FSX doesn't actually support the IDENT; it's just
there to look pretty. The modes dial doesn't actually seem to
do anything either, so at least we didn't waste our time there.

I'm not sure if it's worth trying to bind it to send the "i" key
just for VoxATC---especially since it's re-bindable. I don't want
to have to deal with that.