Refactor event transmission for simplicity
RadioStackModule is no longer a dumb delegate for the Connection!
It is still mostly basic glue, but a lot of code was happily
removed from FsxConnection, so if we ever changed libraries there'd
be much less rewriting. This change also simplifies writing
TestModules, as we no longer have to duplicate the Observers; we
can simply listen for the specific events fired and create an
anonymous Answer to append to the appropriate List. Neat!

This change should make it significantly easier to implement
new equipment (like the Autopilot box!)