Refactor to provide a single stream of SimData objects
This is anticipation of needing more DataType objects, and
continues the previous simplification efforts. Modules can now
provide the appropriate Observables with the
`RxUtil.pickInstancesOf` Operator.

If we ever request events at the "sim frame" rate, the filtering
may become problematic, but we can always add a special case for
such events if it does